use HashRouter for app
hoping to fix the refresh issue because we are using a single page app but are not configuring any server requests. So if user directly goes to a page, the app on the client side would not know how to serve that page. Hash History will make the server just look at the root url and let react-router handle sending the right page